Update NftOrderProtocol and Service
