+++ FakeNFT/Services/NftOrderInfromationService.swift
@@ -1,12 +1,18 @@
 import Foundation
 
 typealias NftOrderCompletion = (Result<NftOrder, Error>) -> Void
+typealias NftOrderPutCompletion = (Result<OrderPutResponse, Error>) -> Void
 
-protocol NftOrderService {
+protocol NftOrderServiceProtocol {
     func loadNftOrder(completion: @escaping NftOrderCompletion)
+    
+    func sendOrderPutRequest(
+        nfts: [String],
+        completion: @escaping NftOrderPutCompletion
+    )
 }
 
-final class NftOrderServiceImpl: NftOrderService {
+final class NftOrderService: NftOrderServiceProtocol {
 
     private let networkClient: NetworkClient
 
@@ -21,6 +27,22 @@ final class NftOrderServiceImpl: NftOrderService {
             switch result {
             case .success(let nftOrder):
                 completion(.success(nftOrder))
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+    }
+    
+    func sendOrderPutRequest(
+        nfts: [String],
+        completion: @escaping NftOrderPutCompletion
+    ) {
+        let dto = NftOrderDtoObject(nfts: nfts)
+        let request = NftOrderPutRequest(dto: dto)
+        networkClient.send(request: request, type: OrderPutResponse.self) { result in
+            switch result {
+            case .success(let putResponse):
+                completion(.success(putResponse))
             case .failure(let error):
                 completion(.failure(error))
             }
